requirements.txt
numpy

tqdm

loguru

grpcio==1.67.1
grpcio-tools==1.67.1

hydra-core==1.3.2
omegaconf==2.3.0